move wire protocol module to tpls directory
