+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/DNSPacket.ts
