+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/constants/DNS_CLASSES.ts
