+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/A_Record.ts
