+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/CNAME_Record.ts
