+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/HINFO_Record.ts
