+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/NS_Record.ts
