+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/PTR_Record.ts
