+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/RDataMap.d.ts
