+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/SOA_Record.ts
