+++ src/modules/transport-layer-subsystem/wire-protocol/DNS-core/resource-records/TXT_Record.ts
