+++ src/modules/transport-layer-subsystem/wire-protocol/WireProtocolModule.ts
@@ -1,9 +1,15 @@
-import type { Module } from '../Module.js';
+import type { Logger } from '../../../logging/Logger.js';
+import type { ConfigManager } from '../../config-module/ConfigModule.js';
+import { Module } from '../../Module.js';
 import { DNSPacket } from './DNS-core/DNSPacket.js';
 import { CursorBuffer } from './parser/CursorBuffer.js';
 import { DNSParser } from './parser/DNSParser.js';
 
-export class WireProtocolModule implements Module {
+export class WireProtocolModule extends Module {
+  constructor(logger: Logger, config: ConfigManager) {
+    super(logger, config);
+  }
+
   decode(buffer: Buffer): DNSPacket {
     const parser = new DNSParser();
