+++ src/modules/transport-layer-subsystem/wire-protocol/parser/CursorBuffer.ts
